% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colclean}
\alias{colclean}
\title{colclean}
\usage{
colclean(tib)
}
\arguments{
\item{tib}{dataframe}
}
\value{
a dataframe with cleaned up variable names
}
\description{
This function makes variable names nicer, in one shot
}
