% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{do}
\alias{do}
\title{Evaluate String as R Code}
\usage{
do(str)
}
\arguments{
\item{str}{A string containing R code to be evaluated.}
}
\value{
The result of evaluating the R code contained in the string.
}
\description{
This function takes a string containing R code and evaluates it.
}
\examples{
do("2 + 2")
do("mean(c(1, 2, 3, 4, 5))")
}
