% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hydrolyze}
\alias{hydrolyze}
\title{Hydrolyze DataFrame Columns to Global Environment}
\usage{
hydrolyze(df)
}
\arguments{
\item{df}{A dataframe to hydrolyze.}
}
\description{
This function takes a dataframe and assigns each column as a vector to the global environment.
}
