% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mashAvocado}
\alias{mashAvocado}
\title{Combine Regression Results}
\usage{
mashAvocado(regs_vec)
}
\arguments{
\item{regs_vec}{A character vector of regression calls.}
}
\value{
A combined dataframe of regression results.
}
\description{
This function takes a vector of regression models and combines their results.
}
