% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name}
\alias{name}
\title{Get Name of the Thing}
\usage{
name(thing)
}
\arguments{
\item{thing}{A variable or expression whose name you want to retrieve.}
}
\value{
A character string representing the name of the variable or expression passed to it.
}
\description{
This function returns the name of the variable (or expression) passed to it as a string.
}
