% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRmd.R
\name{newRmd}
\alias{newRmd}
\title{Create and Open a New R Markdown File}
\usage{
newRmd(start = "")
}
\arguments{
\item{start}{An optional prefix for the filename. If not provided, the filename will just end with ".Rmd".}
}
\value{
Invisible NULL. The function is called for its side effect of creating and opening an R Markdown file.
}
\description{
This function creates a new R Markdown (Rmd) file with an optional prefix and opens it for editing.
}
