% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_functions.R
\name{OLSeq}
\alias{OLSeq}
\title{Generate Equation String from Linear Model}
\usage{
OLSeq(lm)
}
\arguments{
\item{lm}{A linear model object.}
}
\value{
A character string of the model's equation.
}
\description{
Generates a character string representing the equation of the linear model.
}
