% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read}
\alias{read}
\title{Read Data from CSV and Return as Tibble}
\usage{
read(str)
}
\arguments{
\item{str}{The name of the file (without the .csv extension) to be read.}
}
\value{
A tibble containing the data from the CSV file.
}
\description{
This function reads a CSV file into R and returns it as a tibble. 
It expects the filename without the .csv extension.
for reading the file, suppressing any messages generated during the read process.
}
