% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_functions.R
\name{reg_df}
\alias{reg_df}
\title{Regression Summary to DataFrame}
\usage{
reg_df(lm, thin = T)
}
\arguments{
\item{lm}{A linear model object.}

\item{thin}{A logical value indicating whether to return a thinner summary table. 
Default is TRUE.}
}
\value{
A tibble summarizing the regression coefficients and statistics.
}
\description{
Takes a linear model object and returns a data frame summarizing the regression analysis,
including coefficients, standard errors, t values, p-values, and significance levels.
}
