% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_functions.R
\name{regDF}
\alias{regDF}
\title{Regression Summary with Equation and Extremes}
\usage{
regDF(lm)
}
\arguments{
\item{lm}{A linear model object.}
}
\value{
A data frame summarizing key regression outputs and statistics.
}
\description{
Provides a summary of the regression, including the equation, minimum and maximum t values,
and estimates.
}
