% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regstr}
\alias{regstr}
\title{Generate Regression Strings from DataFrame}
\usage{
regstr(df, y_max_ind = 1, usual = TRUE)
}
\arguments{
\item{df}{Dataframe containing the variables.}

\item{y_max_ind}{The index of the last Y variable in the dataframe.}

\item{usual}{Boolean to toggle the return format.}
}
\value{
A character vector of regression formula strings.
}
\description{
This function generates a set of regression formula strings from a given dataframe. 
It's configured to handle dataframes with multiple Y (dependent variables) and X (independent variables).
}
