% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSheets.R
\name{toSheets}
\alias{toSheets}
\title{Open New Google Sheets and Optionally Copy Data to Clipboard}
\usage{
toSheets(x)
}
\arguments{
\item{x}{Optional data to be copied to the clipboard.}
}
\description{
This function opens a new Google Sheets document in the default web browser. 
If data is provided (and is not NA), it attempts to copy this data to the clipboard.
}
