% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vanillaRegs}
\alias{vanillaRegs}
\title{Generate Basic Regression Formulas}
\usage{
vanillaRegs(df, ycount = 1)
}
\arguments{
\item{df}{The dataframe from which to generate regressions.}

\item{ycount}{The number of Y (dependent) variables in the dataframe.}
}
\value{
A character vector of regression calls.
}
\description{
This function generates basic regression formulas from a given dataframe. 
It focuses on creating one predictor at a time regression models.
}
