% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write}
\alias{write}
\title{Write Data Frame or Text to File}
\usage{
write(df = NULL)
}
\arguments{
\item{df}{The data frame or object to be written to a file. Can be NULL.}
}
\value{
None.
}
\description{
This function writes a data frame to a CSV file or any other object to a text file. 
The name of the output file is automatically generated based on the name of the object in R. 
If the object is a data frame, it is saved as a CSV file. Otherwise, it is saved as a text file.
}
